+++ app/src/main/AndroidManifest.xml
@@ -61,10 +61,10 @@
 
         <activity android:name=".core.base.BaseWebviewActivity"></activity>
         <activity android:name=".im.SessionlistActivity"></activity>
-        <activity android:name=".shop.CustomerservicelistActivity" />
-        <activity android:name=".shop.ImActivity"   android:screenOrientation="portrait"
+        <activity android:name=".im.CustomerservicelistActivity" />
+        <activity android:name=".im.ImActivity"   android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan"  />
-        <activity android:name=".shop.Customerservice" />
+        <activity android:name=".im.CustomerserviceActivity" />
         <activity android:name=".travel.TraveFormActivity" />
         <activity android:name=".travel.TraveDetailActivity" /> <!-- 添加provider -->
