+++ app/src/main/java/com/feitianzhu/huangliwo/strategy/request/ListPageRequest.java
@@ -5,6 +5,9 @@ import com.feitianzhu.huangliwo.core.network.BaseRequest;
 import com.feitianzhu.huangliwo.core.network.BaseTravelRequest;
 import com.feitianzhu.huangliwo.core.network.ParamsBuilder;
 import com.feitianzhu.huangliwo.strategy.bean.ListPageBean;
+import com.feitianzhu.huangliwo.utils.StringUtils;
+
+import java.util.List;
 
 import static com.feitianzhu.huangliwo.utils.Urls.TICKET_BASE_URL;
 
@@ -19,8 +22,21 @@ public class ListPageRequest extends BaseTravelRequest {
     //    栏目id 1-会员须知 2-正品保障
     public int columnId;
 
-
-
+    @Override
+    public Object handleRsponseAfterTransform(Object rsp) {
+        if (rsp instanceof ListPageBean) {
+            List<ListPageBean.ListBean> list = ((ListPageBean) rsp).getList();
+            if (list != null && list.size() >= 0) {
+                for (int i = list.size() - 1; i >= 0; i--) {
+                    ListPageBean.ListBean listBean = list.get(i);
+                    if (listBean.getContentType().equals("1") && StringUtils.isEmpty(listBean.getVideo())) {
+                        list.remove(i);
+                    }
+                }
+            }
+        }
+        return super.handleRsponseAfterTransform(rsp);
+    }
     @Override
     public ParamsBuilder appendParams(ParamsBuilder builder) {
         return super.appendParams(builder
